figured that there's no need for a three way merge to count keep visibility of a user, and a buildfix
